+++ classes/No.py
@@ -1,0 +1,8 @@
+class No:
+    def __init__(self, valor):
+        self.valor = valor
+        self.proximo = None
+        self.anterior = None
+
+    def mostra_no(self):
+        print(self.valor)
